+++ wvlet-api/src/main/scala/wvlet/lang/api/Span.scala
@@ -11,10 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package wvlet.lang.compiler.parser
-
-import wvlet.lang.compiler.{Context, CompilationUnit, SourceLocation}
-import wvlet.lang.model.NodeLocation
+package wvlet.lang.api
 
 /**
   * Span is a range between start and end offset, and a point.
