+++ wvlet-api/src/main/scala/wvlet/lang/api/v1/frontend/FrontendApi.scala
@@ -46,10 +46,9 @@ trait FrontendApi:
     * @return
     */
   def fileList(relativePath: String): FileList
-  
+
   def readFile(relativePath: String): String
   def saveFile(relativePath: String, content: String): Unit
-  
 
 object FrontendApi extends RxRouterProvider:
   override def router = RxRouter.of[FrontendApi]
