+++ wvlet-lang/src/main/scala/wvlet/lang/compiler/transform/TrinoRewritePivot.scala
@@ -16,8 +16,8 @@ package wvlet.lang.compiler.transform
 import wvlet.lang.compiler.RewriteRule.PlanRewriter
 import wvlet.lang.compiler.*
 import wvlet.lang.compiler.DBType.Trino
-import wvlet.lang.model.NodeLocation
-import wvlet.lang.compiler.parser.Span.NoSpan
+import wvlet.lang.api.NodeLocation
+import wvlet.lang.api.Span.NoSpan
 import wvlet.lang.model.expr.*
 import wvlet.lang.model.plan.*
