+++ wvlet-lang/src/main/scala/wvlet/lang/model/plan/execution.scala
@@ -13,7 +13,7 @@
  */
 package wvlet.lang.model.plan
 
-import wvlet.lang.model.NodeLocation
+import wvlet.lang.api.NodeLocation
 import wvlet.lang.model.expr.{Attribute, NameExpr}
 import wvlet.airframe.ulid.ULID
