+++ wvlet-lang/src/main/scala/wvlet/lang/model/plan/plan.scala
@@ -14,10 +14,10 @@
 package wvlet.lang.model.plan
 
 import wvlet.lang.catalog.Catalog.TableName
-import wvlet.lang.compiler.parser.Span
-import wvlet.lang.compiler.{SourceFile, TermName, TypeName}
+import wvlet.lang.api.{NodeLocation, Span}
+import wvlet.lang.compiler.{TermName, TypeName, SourceFile}
 import wvlet.lang.model.DataType.{EmptyRelationType, TypeParameter}
-import wvlet.lang.model.{DataType, NodeLocation, RelationType}
+import wvlet.lang.model.{DataType, RelationType}
 import wvlet.lang.model.expr.{Attribute, Expression, NameExpr, QualifiedName, StringLiteral}
 import wvlet.lang.model.plan.LogicalPlan
