+++ wvlet-runner/src/main/scala/wvlet/lang/runner/QueryResult.scala
@@ -14,8 +14,8 @@
 package wvlet.lang.runner
 
 import wvlet.airframe.codec.MessageCodec
-import wvlet.lang.StatusCode
-import wvlet.lang.compiler.SourceLocation
+import wvlet.lang.api.StatusCode
+import wvlet.lang.api.SourceLocation
 import wvlet.lang.model.{DataType, RelationType}
 import wvlet.lang.model.plan.LogicalPlan
 import wvlet.log.LogSupport
